Recommend lv2-bankstown and speakersafetyd
The main purpose of asahi-audio is to make the speakers
in Apple Silicon laptops to sound good.
For this purpose bass enhancement is needed, which is done
via the bankstown plugin for lv2 (lv2-bankstown).

While not strictly part of the scope of making good sound,
speaker protection is also nice to have. Thus recommend speakersafetyd.

Both of these packages are currently not in the debian archive, but
sitting as merge requests for rust-team/debcargo-conf. However since
they are currently only recommends, it's not a problem listing them
now anyway.

Arguably, lv2-bankstown should even be bumped to a dependency
instead of recommends.